feat(knowledgebase-search): Extract KnowledgebaseRegistry into a generic PostRegistry
Done in prep for HelpPostRegistry. A class implementing PostRegistry can determine what content is in the registry
